requirements.txt
MetaTrader5
requests
python-dotenv
pandas
numpy
xgboost
tensorflow
Flask
Flask-Cors
